% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readability.R
\name{plot.readability}
\alias{plot.readability}
\title{Plots a readability Object}
\usage{
\method{plot}{readability}(x, ...)
}
\arguments{
\item{x}{A \code{readability} object.}

\item{\ldots}{ignored.}
}
\description{
Plots a readability object
}

