% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readability.R
\name{print.readability}
\alias{print.readability}
\title{Prints a readability Object}
\usage{
\method{print}{readability}(x, digits = 1, ...)
}
\arguments{
\item{x}{A \code{readability} object.}

\item{digits}{The number of digits to print.}

\item{\ldots}{ignored.}
}
\description{
Prints a readability object
}

